refactor(test): split tests by module under test
Replace monolithic test_code_quality.py with per-module test files:
- test_main.py — argparse contract, pre_pro annotation
- test_json_parser.py — filter mutation, predicates, imports
- test_file_utils.py — return type annotations
- test_package_structure.py — __init__.py existence
